Update Guide.md
Removed a comma that wasn't needed I don't think and changed a "like" to "such as" to not have two "like"s in the same sentence. Added a "-" between "Town Buildings" to make it consistent with "Halloween-Like". This is a separate commit as this one may be more subjective whereas the other was just typo fixes that are definitely correct